+++ ghu.js
@@ -68,7 +68,7 @@ ghu.task('build:scripts', runtime => {
         .then(webpack(webpackCfg([SRC]), {showStats: false}))
         .then(wrap('\n\n// @include "pre.js"\n\n'))
         .then(includeit())
-        .then(ife(() => runtime.args.production, uglify()))
+        .then(ife(() => runtime.args.production, uglify({compressor: {warnings: false}})))
         .then(wrap(runtime.commentJs))
         .then(write(mapper, {overwrite: true}));
 });