+++ src/_h5ai/public/js/lib/ext/preview/preview-txt.js
@@ -1,33 +1,28 @@
 const lolight = require('lolight');
 const marked = require('marked');
-const {each, keys, dom} = require('../../util');
+const {keys, dom} = require('../../util');
 const {win} = require('../../globals');
-const event = require('../../core/event');
 const allsettings = require('../../core/settings');
 const preview = require('./preview');
-const previewX = require('./preview-x');
 
 const XHR = win.XMLHttpRequest;
 const settings = Object.assign({
     enabled: false,
     styles: {}
 }, allsettings['preview-txt']);
-const tplText = '<pre id="pv-content-txt" class="highlighted"></pre>';
-const tplMarkdown = '<div id="pv-content-txt" class="markdown"></div>';
+const tplPre = '<pre id="pv-content-txt"></pre>';
+const tplDiv = '<div id="pv-content-txt"></div>';
 
 let state;
 
-const onAdjustSize = () => {
+const updateGui = () => {
     const el = dom('#pv-content-txt')[0];
     if (!el) {
         return;
     }
 
-    const elContent = dom('#pv-content')[0];
-
-    dom(el).css({
-        height: elContent.offsetHeight - 16 + 'px'
-    });
+    const container = dom('#pv-container')[0];
+    el.style.height = container.offsetHeight - 16 + 'px';
 
     preview.setLabels([
         state.item.label,
@@ -56,43 +51,36 @@ const requestTextContent = href => {
 
 const loadText = item => {
     return requestTextContent(item.absHref)
-        .catch(err => '[ajax error] ' + err)
+        .catch(err => '[request failed] ' + err)
         .then(content => {
-            const style = settings.styles[state.item.type];
-            let $text;
+            const style = settings.styles[item.type];
 
             if (style === 1) {
-                $text = dom(tplText).text(content);
+                return dom(tplPre).text(content);
             } else if (style === 2) {
-                $text = dom(tplMarkdown).html(marked(content));
+                return dom(tplDiv).html(marked(content));
             } else if (style === 3) {
-                $text = dom(tplText);
-                const $code = dom('<code/>').text(content).appTo($text);
+                const $code = dom('<code></code>').text(content);
                 win.setTimeout(() => {
                     lolight.el($code[0]);
                 }, content.length < 20000 ? 0 : 500);
-            } else {
-                $text = dom(tplMarkdown).text(content);
+                return dom(tplPre).app($code);
             }
 
-            return $text;
+            return dom(tplDiv).text(content);
         });
-        // .then(x => new Promise(resolve => setTimeout(() => resolve(x), 1000)));
 };
 
 const onEnter = (items, idx) => {
-    state = previewX.pvState(items, idx, loadText, onAdjustSize);
+    state = preview.state(items, idx, loadText, updateGui);
 };
-
-const initItem = previewX.initItemFn(keys(settings.styles), onEnter);
-const onViewChanged = added => each(added, initItem);
 
 const init = () => {
     if (!settings.enabled) {
         return;
     }
 
-    event.sub('view.changed', onViewChanged);
+    preview.register(keys(settings.styles), onEnter);
 };
 
 init();