+++ src/_h5ai/public/js/lib/ext/preview/preview-vid.js
@@ -1,34 +1,25 @@
-const {each, dom} = require('../../util');
-const event = require('../../core/event');
+const {dom} = require('../../util');
 const allsettings = require('../../core/settings');
 const preview = require('./preview');
-const previewX = require('./preview-x');
 
 const settings = Object.assign({
     enabled: false,
+    autoplay: true,
     types: []
 }, allsettings['preview-vid']);
 const tpl = '<video id="pv-content-vid"/>';
 
 let state;
 
-const onAdjustSize = () => {
+const updateGui = () => {
     const el = dom('#pv-content-vid')[0];
     if (!el) {
         return;
     }
 
-    const elContent = dom('#pv-content')[0];
-    const contentW = elContent.offsetWidth;
-    const contentH = elContent.offsetHeight;
+    preview.centerContent();
+
     const elW = el.offsetWidth;
-    const elH = el.offsetHeight;
-
-    dom(el).css({
-        left: (contentW - elW) * 0.5 + 'px',
-        top: (contentH - elH) * 0.5 + 'px'
-    });
-
     const elVW = el.videoWidth;
     const elVH = el.videoHeight;
 
@@ -43,26 +34,25 @@ const loadVideo = item => {
     return new Promise(resolve => {
         const $el = dom(tpl)
             .on('loadedmetadata', () => resolve($el))
-            // .attr('autoplay', 'autoplay')
             .attr('controls', 'controls')
             .attr('src', item.absHref);
+
+        if (settings.autoplay) {
+            $el.attr('autoplay', 'autoplay');
+        }
     });
-    // .then(x => new Promise(resolve => setTimeout(() => resolve(x), 1000)));
 };
 
 const onEnter = (items, idx) => {
-    state = previewX.pvState(items, idx, loadVideo, onAdjustSize);
+    state = preview.state(items, idx, loadVideo, updateGui);
 };
-
-const initItem = previewX.initItemFn(settings.types, onEnter);
-const onViewChanged = added => each(added, initItem);
 
 const init = () => {
     if (!settings.enabled) {
         return;
     }
 
-    event.sub('view.changed', onViewChanged);
+    preview.register(settings.types, onEnter);
 };
 
 init();